backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter~=2.4.0
djoser==2.2.3
Pillow
reportlab
isort==5.13.2
python-dotenv==1.0.1
gunicorn==20.1.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
python-decouple
flake8
flake8-isort
